backend/pyproject.toml
[project]
name = "mocotch-backend"
version = "0.1.0"
description = "RPG制作・実行ツール バックエンド"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
    "gitpython>=3.1.43",
    "python-multipart>=0.0.17",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
